requirements.txt
Flask
Flask-Cors
SQLAlchemy
praw
python-dotenv
spacy==2.3.1
